ai_backend/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
google-generativeai
numpy
python-dotenv
